+++ src/Abp/Runtime/Session/IAbpSession.cs
@@ -37,7 +37,7 @@ namespace Abp.Runtime.Session
         /// TenantId of the impersonator.
         /// This is filled if a user with <see cref="ImpersonatorUserId"/> performing actions behalf of the <see cref="UserId"/>.
         /// </summary>
-        int? ImpersonatorTenantId { get; }
+        long? ImpersonatorTenantId { get; }
 
         /// <summary>
         /// Used to change <see cref="TenantId"/> and <see cref="UserId"/> for a limited scope.
